Fix connection issue in Amazon Keyspaces
In Amazon Keyspaces, system.local returns the localhost address and
system.peers returns a host that contains the same hostID as localhost.
This causes the connection issue in refreshRing(), where host address is
overwritten with the localhost address and a host with the same hostID
results in the "cannot find host" error.

This commit fixes the issue by making map from a slice of hosts.